post-processing-filters extracted to own file
