+++ common/viewer.h
@@ -7,6 +7,7 @@
 #include "model-views.h"
 #include "subdevice-model.h"
 #include "stream-model.h"
+#include "post-processing-filters.h"
 #include "notifications.h"
 #include "skybox.h"
 #include "measurement.h"
